put delete modal in separate file
put delete modal in separate file
