+++ src/pages/admin/coordinator/index.tsx
@@ -1,7 +1,7 @@
 import Layout from 'components/admin/Layout'
 import { DataGrid, GridColumns, GridRowId } from '@mui/x-data-grid'
 import { useCoordinatorList } from 'common/hooks/coordinators'
-import { Box, Button, IconButton, Modal, Typography } from '@mui/material'
+import { Button, IconButton, Modal, Typography } from '@mui/material'
 import DeleteIcon from '@mui/icons-material/Delete'
 import EditIcon from '@mui/icons-material/Edit'
 import PageviewIcon from '@mui/icons-material/Pageview'
@@ -17,25 +17,15 @@ import { ApiErrors } from 'common/api-errors'
 import { AxiosError, AxiosResponse } from 'axios'
 import router from 'next/router'
 import ViewModal from 'components/admin/coordinator/ViewModal'
-
-const style = {
-  position: 'absolute' as const,
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 400,
-  bgcolor: 'background.paper',
-  boxShadow: 24,
-  p: 4,
-}
+import DeleteModal from 'components/admin/coordinator/DeleteModal'
 
 function Coordinator() {
   const { data = [] } = useCoordinatorList()
   const [selectedRows, setSelectedRows] = React.useState<CoordinatorResponse[]>([])
   const [selectionModel, setSelectionModel] = React.useState<GridRowId[]>([])
   const [isDeleteSelectedModalOpen, setIsDeleteSelectedModalOpen] = React.useState(false)
-  const [open, setOpen] = React.useState(false)
-  const [row, setRow] = React.useState<CoordinatorResponse>()
+  const [isViewModalOpen, setIsViewModalOpen] = React.useState(false)
+  const [selectedRow, setSelectedRow] = React.useState<CoordinatorResponse>()
   const [rowToDelete, setRowToDelete] = React.useState<{ id: string }>()
   const [isDeleteModalOpen, setIsDeleteModalOpen] = React.useState(false)
 
@@ -47,12 +37,12 @@ function Coordinator() {
 
   const handleOpen = (row: any) => {
     return () => {
-      setOpen(true)
-      setRow(row)
+      setIsViewModalOpen(true)
+      setSelectedRow(row)
     }
   }
   const handleClose = () => {
-    setOpen(false)
+    setIsViewModalOpen(false)
   }
 
   const handleDeleteModalOpen = (row: any) => {
@@ -129,6 +119,7 @@ function Coordinator() {
     setSelectedRows([])
     setSelectionModel([])
   }
+
   return (
     <Layout>
       <h1>All coordinators</h1>
@@ -161,26 +152,11 @@ function Coordinator() {
         isOpen={isDeleteSelectedModalOpen}
         deleteHandler={deleteSelectedHandler}
         handleDeleteModalClose={closeDeleteSelectedHandler}></DeleteSelectedModal>
-      <ViewModal open={open} handleClose={handleClose} data={row}></ViewModal>
-      <Modal
+      <ViewModal open={isViewModalOpen} handleClose={handleClose} data={selectedRow}></ViewModal>
+      <DeleteModal
         open={isDeleteModalOpen}
-        onClose={handleDeleteModalClose}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description">
-        <Box sx={style}>
-          <Typography id="modal-modal-title" variant="h6" component="h2">
-            Delete?
-          </Typography>
-          <Typography id="modal-modal-description" sx={{ mt: 2 }}>
-            <Button variant="contained" onClick={deleteHandler}>
-              Yes
-            </Button>
-            <Button variant="contained" onClick={handleDeleteModalClose}>
-              No
-            </Button>
-          </Typography>
-        </Box>
-      </Modal>
+        handleClose={handleDeleteModalClose}
+        deleteHandler={deleteHandler}></DeleteModal>
     </Layout>
   )
 }
